Clean up, update links, new recommendations for Safari/iOS
We link to Better Blocker when Safari is detected in this version as uBlock Origin for Safari seems not to be happening because of changes Apple made to restrict ad-blocking on Safari. We've also removed Mozilla Focus as the recommendation when iOS is detected, pointing visitors to Better Blocker instead. Mozilla Focus installed content blockers too, but it was also its own browser. We want this to lead to a simple action for the user, so no recommending a different browser (although perhaps we should present one better suited for ad blocking as an option).